change routing structure and hosting react app
